requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.20.51
botocore==1.23.51
Brotli==1.0.9
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
celery==5.2.6
certifi==2021.5.30
cffi==1.14.5
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.4
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.2
constantly==15.1.0
cryptography==37.0.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.4
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
EditorConfig==0.12.3
firebase-admin==5.2.0
Flask==2.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
funcy==1.16
gevent==21.8.0
geventhttpclient==1.5.1
google-api-core==1.31.2
google-api-python-client==2.17.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.2.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
h11==0.13.0
hiredis==2.0.0
httpie==2.6.0
httplib2==0.19.1
httptools==0.4.0
hyperlink==21.0.0
iamport-rest-client==0.9.0
idna==3.2
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsbeautifier==1.14.0
kombu==5.2.4
MarkupSafe==2.0.1
msgpack==1.0.2
mysqlclient==2.1.0
oauthlib==3.2.2
packaging==21.0
phonenumbers==8.12.27
Pillow==8.3.1
prompt-toolkit==3.0.29
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.2.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
roundrobin==0.0.2
rsa==4.7.2
s3transfer==0.5.0
schedule==1.1.0
sentry-sdk==1.11.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.13
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websocket==0.2.1
websockets==10.3
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0